Fixing nodejs link creation if the file already exists.

+++ recipes/utilities.rb
@@ -8,6 +8,7 @@ include_recipe 'composer'
 execute 'node old executable linking' do
   command 'sudo ln -s /usr/bin/nodejs /usr/bin/node'
   only_if { platform_family?('debian') }
+  not_if { File.exist?('/usr/bin/node') }
 end
 
 nodejs_npm 'grunt'
